Move HexComparere to Hex::Comparer service
