+++ app/decorators/organization_decorator.rb
@@ -1,6 +1,6 @@
 class OrganizationDecorator < ApplicationDecorator
   def darker_color(adjustment = 0.88)
-    HexComparer.new([enriched_colors[:bg], enriched_colors[:text]]).brightness(adjustment)
+    Hex::Comparer.new([enriched_colors[:bg], enriched_colors[:text]]).brightness(adjustment)
   end
 
   def enriched_colors
