+++ app/decorators/user_decorator.rb
@@ -32,7 +32,7 @@ class UserDecorator < ApplicationDecorator
   end
 
   def darker_color(adjustment = 0.88)
-    HexComparer.new([enriched_colors[:bg], enriched_colors[:text]]).brightness(adjustment)
+    Hex::Comparer.new([enriched_colors[:bg], enriched_colors[:text]]).brightness(adjustment)
   end
 
   def enriched_colors
